New custom Exception message

+++ src/main/java/com/malise/app/service/product/commandhandlers/DeleteProductCommandHandler.java
@@ -1,9 +1,12 @@
 package com.malise.app.service.product.commandhandlers;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.malise.app.Exceptions.ProductNotFoundException;
 import com.malise.app.model.product.Product;
 import com.malise.app.repository.product.ProductRepository;
 import com.malise.app.service.product.Command;
@@ -16,6 +19,12 @@ public class DeleteProductCommandHandler implements Command<Integer, ResponseEnt
 
   @Override
   public ResponseEntity<ResponseEntity> execute(Integer id) {
+
+    Optional<Product> optionalProduct = productRepository.findById(id);
+
+    if (optionalProduct.isEmpty()) {
+      throw new ProductNotFoundException();
+    }
 
     Product product = productRepository.findById(id).get();
     productRepository.delete(product);
